move useFetch to custom hooks
